backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.23
alembic>=1.7.3
python-dotenv>=0.19.0
celery>=5.1.2
redis>=3.5.3
psycopg2-binary>=2.9.1